Don't get confused by back references referring to the register named "NIL".
This fixes issue #12.  Back references referring to the register named
"NIL" were being associated with the register with name NIL (the
symbol) instead of "NIL" (the string).  In this case NIL (symbol) is
supposed to indicate that the register is unnamed, but this was not
being checked with STRINGP before comparison with STRING=.