Don't create a separate matcher for matching registers from subpattern references.
Instead, every time we descend into a register from a subpattern
reference, we first push a value onto a register-specific list and pop
it off once we return.  When STORE-END-OF-REG sees that there is a
value on the list, it knows we entered the register from a subpattern
reference and doesn't try to match the part of the regex following the
register.

It's hard to say whether this improves or degrades performance and
readability.  But it does seem simpler.