Don't store possible register offset of register entered via subpattern reference.
This was causing many tests to fail with bizarre error messages.  It
wasn't caught using the old register offsets storage model since the
array the value was beging stored in was a temporary array to begin
with.

With this commit, the only tests run by RUN-ALL-TESTS that fail are
those that Perl itself gets wrong.